pyproject.toml
[project]
name = "strava"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "humanize>=4.14.0",
    "pydantic>=2.12.4",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.5",
]

[dependency-groups]
qa = [
    "black>=25.11.0",
    "flake8>=7.3.0",
    "isort>=7.0.0",
    "mypy>=1.18.2",
    "pytest>=9.0.1",
    "types-python-dateutil>=2.9.0.20251115",
    "types-requests>=2.32.4.20250913",
]
